GP-2448: Sync DebuggerModelProvider to tool

+++ Ghidra/Debug/Debugger/src/main/java/ghidra/app/plugin/core/debug/gui/time/DebuggerTimeProvider.java
@@ -117,7 +117,7 @@ public class DebuggerTimeProvider extends ComponentProviderAdapter {
 				myActionContext = null;
 				return;
 			}
-			if (snap.longValue() == current.getSnap().longValue()) {
+			if (snap.longValue() == current.getSnap()) {
 				return;
 			}
 			myActionContext = new DebuggerSnapActionContext(current.getTrace(), snap);
